Refactor k-view
